Fixed bug in senseLead miner code
There was code in senseLead's miner that was
misuing senseLead. senseLead returns an int
(the amount of lead) and it was be used as if
it returned a Boolean like canMineLead. I added
a value check to senseLead to make it a Boolean
expression. There are comments in the code about
some tactics surrounding that code.